Add files via upload
merged the functionality of both branches. scroll on participant lists needs to be fixed though, links to settings from main page need to be done, dialpads need to reset when call button is pressed. additional images can be added to video call as mock cameras. speaking icons need to be added to all calls.